Добавил закрывающий тег div.page-wrapper и разложил php по файлам
