+++ config.php
@@ -1,0 +1,4 @@
+<?php
+
+$is_auth = rand(0, 1);
+$user_name = 'Валерий'; // укажите здесь ваше имя
