+++ index.php
@@ -1,83 +1,8 @@
 <?php
 
-$is_auth = rand(0, 1);
-
-$category_list = [
-    "Доски и лыжи",
-    "Крепления",
-    "Ботинки",
-    "Одежда",
-    "Инструменты",
-    "Разное"
-];
-
-$advertising_list = [
-    [
-        "name" => "2014 Rossignol District Snowboard",
-        "category" => "Доски и лыжи",
-        "price" => "10999",
-        "image" => "img/lot-1.jpg"
-    ],
-    [
-        "name" => "DC Ply Mens 2016/2017 Snowboard",
-        "category" => "Доски и лыжи",
-        "price" => "159999",
-        "image" => "img/lot-2.jpg"
-    ],
-    [
-        "name" => "Крепления Union Contact Pro 2015 года размер L/XL",
-        "category" => "Крепления",
-        "price" => "8000",
-        "image" => "img/lot-3.jpg"
-    ],
-    [
-        "name" => "Ботинки для сноуборда DC Mutiny Charocal",
-        "category" => "Ботинки",
-        "price" => "10999",
-        "image" => "img/lot-4.jpg"
-    ],
-    [
-        "name" => "Куртка для сноуборда DC Mutiny Charocal",
-        "category" => "Одежда",
-        "price" => "7500",
-        "image" => "img/lot-5.jpg"
-    ],
-    [
-        "name" => "Маска Oakley Canopy",
-        "category" => "Разное",
-        "price" => "5400",
-        "image" => "img/lot-6.jpg"
-    ],
-];
-
-$user_name = 'Валерий'; // укажите здесь ваше имя
-
-function price_format($cost) {
-    $cost_ceil = ceil($cost);
-    if ($cost_ceil >= 1000) {
-        $cost_format = number_format($cost_ceil, 0, ',', ' ');
-    } else {
-        $cost_format = $cost_ceil;
-    }
-    return ($cost_format . ' <b class="rub">р</b>');
-};
-
-function include_template($name, array $data = []) {
-    $name = 'templates/' . $name;
-    $result = '';
-
-    if (!is_readable($name)) {
-        return $result;
-    }
-
-    ob_start();
-    extract($data);
-    require $name;
-
-    $result = ob_get_clean();
-
-    return $result;
-}
+require_once('config.php');
+require_once('functions.php');
+require_once('data.php');
 
 $main_content = include_template('main.php', [
     'category_list' => $category_list,
